multiple flows implementation
