+++ pom.xml
@@ -78,7 +78,7 @@
                             <transformers>
                                 <!-- Include the main class in the manifest -->
                                 <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                                    <mainClass>com.bibledetector.BibleVerseDetector</mainClass>
+                                    <mainClass>com.bibledetector.BibleDetector</mainClass>
                                 </transformer>
                             </transformers>
                         </configuration>
