--- src/main/java/com/bibledetector/steps/Actions.java
@@ -1,4 +1,0 @@
-package com.bibledetector.steps;
-
-public interface Actions extends Step {
-}
