--- src/main/java/com/bibledetector/steps/Execution.java
@@ -1,4 +1,0 @@
-package com.bibledetector.steps;
-
-public interface Execution extends Step {
-}
