--- src/main/java/com/bibledetector/steps/Extraction.java
@@ -1,4 +1,0 @@
-package com.bibledetector.steps;
-
-public interface Extraction extends Step {
-}
