--- src/main/java/com/bibledetector/steps/Selection.java
@@ -1,4 +1,0 @@
-package com.bibledetector.steps;
-
-public interface Selection extends Step {
-}
