--- src/main/java/com/bibledetector/steps/Validation.java
@@ -1,4 +1,0 @@
-package com.bibledetector.steps;
-
-public interface Validation extends Step {
-}
