+++ src/main/java/com/bibledetector/steps/bibleflow/Selection.java
@@ -1,0 +1,6 @@
+package com.bibledetector.steps.bibleflow;
+
+import com.bibledetector.steps.types.Step;
+
+public interface Selection extends Step {
+}
