+++ src/main/java/com/bibledetector/steps/bibleflow/impl/actions/GoLivePRS.java
@@ -1,0 +1,11 @@
+package com.bibledetector.steps.bibleflow.impl.actions;
+
+import com.bibledetector.steps.bibleflow.types.impl.ActionInput;
+import com.bibledetector.steps.bibleflow.types.step.Action;
+
+public class GoLivePRS implements Action {
+    @Override
+    public void execute(ActionInput actionInput) {
+        // TODO: Integrate with PRS
+    }
+}
