+++ src/main/java/com/bibledetector/steps/bibleflow/types/impl/ActionResult.java
@@ -1,0 +1,11 @@
+package com.bibledetector.steps.bibleflow.types.impl;
+
+import com.bibledetector.constants.ResultStatusCode;
+import com.bibledetector.steps.types.StepResult;
+
+public record ActionResult(String book, String chapter, String verse, ResultStatusCode statusCode) implements StepResult {
+    @Override
+    public ResultStatusCode resultStatusCode() {
+        return statusCode;
+    }
+}
