+++ src/main/java/com/bibledetector/steps/bibleflow/types/impl/SelectionInput.java
@@ -1,0 +1,6 @@
+package com.bibledetector.steps.bibleflow.types.impl;
+
+import com.bibledetector.steps.types.StepInput;
+
+public record SelectionInput(String book, String chapter, String verse) implements StepInput {
+}
