+++ src/main/java/com/bibledetector/steps/bibleflow/types/step/Action.java
@@ -1,0 +1,7 @@
+package com.bibledetector.steps.bibleflow.types.step;
+
+import com.bibledetector.steps.bibleflow.types.impl.ActionInput;
+
+public interface Action {
+    void execute(ActionInput actionInput);
+}
