--- src/main/java/com/bibledetector/steps/impl/actions/GoLivePRS.java
@@ -1,11 +1,0 @@
-package com.bibledetector.steps.impl.actions;
-
-import com.bibledetector.steps.types.impl.ActionInput;
-import com.bibledetector.steps.types.step.Action;
-
-public class GoLivePRS implements Action {
-    @Override
-    public void execute(ActionInput actionInput) {
-        // TODO: Integrate with PRS
-    }
-}
