+++ src/main/java/com/bibledetector/steps/types/SpeechRecognitionOutput.java
@@ -1,0 +1,4 @@
+package com.bibledetector.steps.types;
+
+public record SpeechRecognitionOutput(String text) implements StepInput {
+}
