+++ src/main/java/com/bibledetector/steps/types/Step.java
@@ -1,0 +1,7 @@
+package com.bibledetector.steps.types;
+
+public interface Step {
+    StepResult run(StepInput stepInput);
+
+    StepInput extractParam(StepResult stepResult);
+}
