--- src/main/java/com/bibledetector/steps/types/impl/ExecutionInput.java
@@ -1,6 +1,0 @@
-package com.bibledetector.steps.types.impl;
-
-import com.bibledetector.steps.types.StepInput;
-
-public record ExecutionInput(String book, String chapter, String verse) implements StepInput {
-}
