--- src/main/java/com/bibledetector/steps/types/impl/ExtractionInput.java
@@ -1,6 +1,0 @@
-package com.bibledetector.steps.types.impl;
-
-import com.bibledetector.steps.types.StepInput;
-
-public record ExtractionInput(String text) implements StepInput {
-}
