--- src/main/java/com/bibledetector/steps/types/impl/ValidationResult.java
@@ -1,11 +1,0 @@
-package com.bibledetector.steps.types.impl;
-
-import com.bibledetector.constants.ResultStatusCode;
-import com.bibledetector.steps.types.StepResult;
-
-public record ValidationResult(String book, String chapter, String verse, ResultStatusCode statusCode) implements StepResult {
-    @Override
-    public ResultStatusCode resultStatusCode() {
-        return statusCode;
-    }
-}
