--- src/main/java/com/bibledetector/steps/types/step/Action.java
@@ -1,7 +1,0 @@
-package com.bibledetector.steps.types.step;
-
-import com.bibledetector.steps.types.impl.ActionInput;
-
-public interface Action {
-    void execute(ActionInput actionInput);
-}
